+++ cmd/function_new.go
@@ -75,33 +75,41 @@ func validateFunctionName(functionName string) error {
 }
 
 func runFunctionNew(cmd *cobra.Command, args []string) error {
-	fmt.Println(checkDefaultMessage)
-	if err := get.CheckFaasCli(); err != nil {
+	verboseFlag, _ := cmd.Flags().GetBool("verbose")
+	exist, err := get.CheckBinary(faasBinaryName, false, verboseFlag)
+	if err != nil {
 		return err
+	}
+
+	if !exist {
+		return fmt.Errorf(notExistBinary)
 	}
 
 	faasCliPath := config.GetFaasCliPath()
 	flagLang, _ := cmd.Flags().GetString("lang")
+	functionName := args[0]
 
-	if err := downloadTemplate(faasCliPath, flagLang); err != nil {
+	if err := downloadTemplate(faasCliPath, flagLang, verboseFlag); err != nil {
 		return err
 	}
 	
-	if err := createFunction(faasCliPath, args[0], flagLang); err != nil {
+	if err := createFunction(faasCliPath, functionName, flagLang, verboseFlag); err != nil {
 		return err
 	}
-	
+
+	fmt.Println("\nFunction " + functionName + " criada com sucesso!")
+
 	return nil
 }
 
-func downloadTemplate(faasCliPath, lang string) error {
-	setStore()
+func downloadTemplate(faasCliPath, lang string, verboseFlag bool) error {
 	taskDownloadTemplate := execute.Task{
 		Command:     faasCliPath,
 		Args:        []string{
-			"template", "store", "pull", lang,
+			"template", "store", "pull", lang, "--url", faasTemplateStoreURL,
 		},
-		StreamStdio: true,
+		StreamStdio:  verboseFlag,
+		PrintCommand: verboseFlag,
 	}
 
 	fmt.Println("Baixando template...")
@@ -111,39 +119,40 @@ func downloadTemplate(faasCliPath, lang string) error {
 	}
 
 	if res.ExitCode > 1 {
-		return fmt.Errorf("exit code %d", res.ExitCode)
+		return fmt.Errorf(res.Stderr)
 	}
 
 	return nil
 }
 
-func createFunction(faasCliPath, projectName, lang string) error {
+func createFunction(faasCliPath, functionName, lang string, verboseFlag bool) error {
 	taskCreateFunction := execute.Task{
 		Command:     faasCliPath,
 		Args:        []string{
-			"new", projectName,
+			"new", functionName,
 			"--lang", lang,
 			"--gateway", "http://gateway.ipaas.localdomain:8080",
 			"--prefix", "registry.localdomain:5000",
 		},
-		StreamStdio: true,
+		StreamStdio:  verboseFlag,
+		PrintCommand: verboseFlag,
 	}
 
-	fmt.Println("Criando template...")
+	fmt.Println("Criando function " + functionName + "...")
 	res, err := taskCreateFunction.Execute()
 	if err != nil {
 		return err
 	}
 
 	if res.ExitCode != 0 {
-		return fmt.Errorf("exit code %d", res.ExitCode)
+		return fmt.Errorf(res.Stderr)
 	}
 
 	if err := writeGitignore(); err != nil {
 		return err
 	}
 
-	if err := addFileEnv(projectName); err != nil {
+	if err := addFileEnv(functionName); err != nil {
 		return err
 	}