+++ pkg/get/check.go
@@ -6,53 +6,54 @@ import (
 	"os/exec"
 )
 
-var errorCheck = "não foi possível baixar o pacote "
+func CheckBinary(binary string, downloadIfNotExist, verboseFlag bool) (bool, error) {
+	exists, _ := existsBinary(binary)
+	errorCheck := "não foi possível baixar o pacote: "
 
-func CheckKubectl() error {
-	if exists, _ := ExistsBinary("kubectl"); !exists {
-		fmt.Println("Baixando kubectl...")
-		if err := DownloadKubectl(); err != nil {
-			return fmt.Errorf(errorCheck + "kubectl")
+	if verboseFlag {
+		fmt.Println("[+] Verificando " + binary)
+	}
+
+	if !exists && downloadIfNotExist {
+		fmt.Println("Baixando " + binary + "...")
+		switch binary {
+		case "kubectl":
+			if err := DownloadKubectl(); err != nil {
+				if verboseFlag {
+					fmt.Println(err)
+				}
+				return false, fmt.Errorf(errorCheck + "kubectl. Tente novamente")
+			}
+		case "k3d":
+			if err := DownloadK3d(); err != nil {
+				if verboseFlag {
+					fmt.Println(err)
+				}
+				return false, fmt.Errorf(errorCheck + "k3d. Tente novamente")
+			}
+		case "helm":
+			if err := DownloadHelm(); err != nil {
+				if verboseFlag {
+					fmt.Println(err)
+				}
+				return false, fmt.Errorf(errorCheck + "helm. Tente novamente")
+			}
+		case "faas-cli":
+			if err := DownloadFaasCli(); err != nil {
+				if verboseFlag {
+					fmt.Println(err)
+				}
+				return false, fmt.Errorf(errorCheck + "faas-cli. Tente novamente")
+			}
+		default:
+			return false, fmt.Errorf("nome de binário inválido")
 		}
 	}
 
-	return nil
+	return exists, nil
 }
 
-func CheckK3d() error {
-	if exists, _ := ExistsBinary("k3d"); !exists {
-		fmt.Println("Baixando k3d...")
-		if err := DownloadK3d(); err != nil {
-			return fmt.Errorf(errorCheck + "k3d")
-		}
-	}
-
-	return nil
-}
-
-func CheckHelm() error {
-	if exists, _ := ExistsBinary("helm"); !exists {
-		fmt.Println("Baixando helm...")
-		if err := DownloadHelm(); err != nil {
-			return fmt.Errorf(errorCheck + "helm")
-		}
-	}
-
-	return nil
-}
-
-func CheckFaasCli() error {
-	if exists, _ := ExistsBinary("faas-cli"); !exists {
-		fmt.Println("Baixando faas-cli...")
-		if err := DownloadFaasCli(); err != nil {
-			return fmt.Errorf(errorCheck + "faas-cli")
-		}
-	}
-
-	return nil
-}
-
-func ExistsBinary(binary string) (exists bool, err error) {
+func existsBinary(binary string) (exists bool, err error) {
 	path, err := exec.LookPath(fmt.Sprintf("%sbin/%s", config.GetUserDir(), binary))
 	exists = path != ""
 	return